+++ CHANGELOG.md
@@ -4,7 +4,7 @@
  * Update build for FFmpeg on Raspbian Stretch ([pull #548](https://github.com/bytedeco/javacpp-presets/pull/548))
  * Add presets for TensorRT
  * Fix build for FFmpeg on `android-x86` and `android-x86_64` platforms ([issue bytedeco/javacv#945](https://github.com/bytedeco/javacv/issues/945))
- * Upgrade presets for HDF5 1.10.2, LLVM 6.0.0, TensorFlow 1.7.0
+ * Upgrade presets for HDF5 1.10.2, LLVM 6.0.0, Tesseract 4.0.0-beta.1 ([issue #385](https://github.com/bytedeco/javacpp-presets/issues/385)), TensorFlow 1.7.0
 
 ### March 29, 2018 version 1.4.1
  * Disable unneeded error messages from LibTIFF in presets for Leptonica ([issue deeplearning4j/DataVec#518](https://github.com/deeplearning4j/DataVec/pull/518))